refactor(web/board): move to board/components/, use shared constants and utils
