+++ apps/web/src/components/board/components/Board.tsx
@@ -1,4 +1,5 @@
-import { type WorkItemDto, fetchIssues, fetchMilestoneIssues } from '@/lib/api';
+import { fetchIssues, fetchMilestoneIssues } from '@/lib/api';
+import type { WorkItemDto } from '@/lib/types';
 import { LANES, laneForState, sortLaneItems } from '@/lib/lanes.config';
 import { useActiveMilestone } from '@/state/active-milestone';
 import { useLaneVisibility } from '@/state/lane-visibility';
