--- apps/web/src/components/board/placeholders.ts
@@ -1,2 +1,0 @@
-/** Placeholder shown in the cost slot on issue cards until M9 wires real data. */
-export const COST_PLACEHOLDER = '$—';
